# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Descripcion del API para el submodulo de rutas",
    "version": "1.0.0",
    "title": "Ms-rutas"
  },
  "host": "100.26.181.55:80",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/noderoute/createRoute": {
      "post": {
        "tags": [
          "Nodo Ruta"
        ],
        "summary": "Agregar una nueva ruta en la base de datos",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto de ruta que va a ser guardado",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idRoute": {
                  "type": "string"
                },
                "startCity": {
                  "type": "string"
                },
                "arrivalCity": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "latitudeStart": {
                  "type": "string"
                },
                "longitudeStart": {
                  "type": "string"
                },
                "latitudeEnd": {
                  "type": "string"
                },
                "longitudeEnd": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "idNodeRoute": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
...
    "/noderoute/getRouteById": {
      "post": {
        "tags": [
          "Nodo Ruta"
        ],
        "summary": "Consulta una ruta con su id",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto de ruta que va a ser consultado",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idRoute": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "idRoute": {
                  "type": "string"
                },
                "startCity": {
                  "type": "string"
                },
                "arrivalCity": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "latitudeStart": {
                  "type": "string"
                },
                "longitudeStart": {
                  "type": "string"
                },
                "latitudeEnd": {
                  "type": "string"
                },
                "longitudeEnd": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/noderoute/getRouteIdById": {
      "post": {
        "tags": [
          "Nodo Ruta"
        ],
        "summary": "Consulta una ruta con su id",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto de ruta que va a ser consultado",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idRoute": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "idNodeRoute": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
    "/noderoute/modifyRoute": {
      "put": {
        "tags": [
          "Nodo Ruta"
        ],
        "summary": "Modificar una ruta con el id",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto de ruta que va a ser modificado",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idRoute": {
                  "type": "string"
                },
                "startCity": {
                  "type": "string"
                },
                "arrivalCity": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "latitudeStart": {
                  "type": "string"
                },
                "longitudeStart": {
                  "type": "string"
                },
                "latitudeEnd": {
                  "type": "string"
                },
                "longitudeEnd": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}